requirements.txt
aioredis==2.0.0
aiosmtplib==1.1.6
anyio==3.3.4
asgiref==3.4.1
async-timeout==3.0.1
bcrypt==3.2.0
blinker==1.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==8.0.3
dnspython==2.1.0
email-validator==1.1.3
fakeredis==1.6.1
fastapi==0.68.2
fastapi-mail==1.0.1
greenlet==1.1.2
h11==0.12.0
httpcore==0.13.7
httpx==0.19.0
idna==3.3
Jinja2==3.0.2
MarkupSafe==2.0.1
packaging==21.0
psycopg2==2.9.1
py-dotenv==0.1
pycparser==2.20
pydantic==1.8.2
pyparsing==3.0.1
python-dotenv==0.19.1
python-multipart==0.0.5
redis==3.5.3
rfc3986==1.5.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
SQLAlchemy==1.4.26
starlette==0.14.2
typing-extensions==3.10.0.2
uvicorn==0.15.0
